Write first update entry and start code comments for update process
